add book depth

+++ rust_native/src/book_info.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone)]
 pub struct BookInfo {
     pub title: String,
-    pub width: f32, // unit is cm
+    pub width: u16, // unit is mm
 }
